transit assignment updates for v9 network

+++ tm2py/components/network/transit/transit_network.py
@@ -507,6 +507,8 @@ class ApplyFares(Component):
                     fs_data["FAREFROMFS"] = [float(x) for x in fs_data["FAREFROMFS"].split(",")]
                 if fs_data["STRUCTURE"] == "FLAT":
                     fs_data["IBOARDFARE"] = float(fs_data["IBOARDFARE"])
+                    # fs_data["IBOARDFARE"] = round(float(fs_data["IBOARDFARE"])*2)/2
+                    # fs_data["FAREFROMFS"] = [round(float(x)*2)/2 for x in fs_data["FAREFROMFS"]]
                 elif fs_data["STRUCTURE"] == "FROMTO":
                     fmi, one, farematrix_id = fs_data["FAREMATRIX"].split(".")
                     fs_data["FAREMATRIX ID"] = int(farematrix_id)
@@ -910,7 +912,11 @@ class ApplyFares(Component):
                     for fs_id, fare1 in xfer_fares1.items():
                         fare2 = xfer_fares2[fs_id]
                         if fare1 != fare2 and (fare1 != "TOO_FAR" and fare2 != "TOO_FAR"):
-                            return False
+                            # if the difference between two fares are less that a number, then treat them as the same fare
+                            if isinstance(fare1, float) and isinstance(fare2, float) and (abs(fare1 - fare2)<=1.5):
+                                continue
+                            else:
+                                return False
             return True
 
         # group faresystems together which have the same transfer-to pattern, 
@@ -959,7 +965,13 @@ class ApplyFares(Component):
             xfer_fares = {}
             for fs_id in faresystems.keys():
                 to_fares = [f[fs_id] for f in xfer_fares_list if f[fs_id] != "TOO_FAR"]
-                fare = to_fares[0] if len(to_fares) > 0 else 0.0
+                # fare = to_fares[0] if len(to_fares) > 0 else 0.0
+                if len(to_fares) == 0:
+                    fare = 0.0
+                elif all(isinstance(item, float) for item in to_fares): # caculate the average here becasue of the edits in matching_xfer_fares function
+                    fare = round(sum(to_fares)/len(to_fares),2)
+                else:
+                    fare = to_fares[0]
                 xfer_fares[fs_id] = fare
             faresystem_groups.append((group, xfer_fares))
             for fs_id in group:
@@ -981,6 +993,7 @@ class ApplyFares(Component):
         self.create_attribute("TRANSIT_LINE", "#src_veh", self.scenario, network, "STRING")
 
         transit_modes = set([m for m in network.modes() if m.type == "TRANSIT"])
+        transit_modes -= set([m for m in network.modes() if m.description == "pnrdummy"]) #remove PNR dummy route from transit modes
         mode_desc = {m.id: m.description for m in transit_modes}
         get_mode_id = network.available_mode_identifier
         get_vehicle_id = network.available_transit_vehicle_identifier
@@ -992,22 +1005,24 @@ class ApplyFares(Component):
                 link.modes |= set([meta_mode])
         lines = _defaultdict(lambda: [])
         for line in network.transit_lines():
-            lines[line.vehicle.id].append(line)
+            if line.mode.id != "p": #remove PNR dummy route from transit modes
+                lines[line.vehicle.id].append(line)
             line["#src_mode"] = line.mode.id
             line["#src_veh"] = line.vehicle.id
         for vehicle in network.transit_vehicles():
-            temp_veh = network.create_transit_vehicle(get_vehicle_id(), vehicle.mode.id)
-            veh_id = vehicle.id
-            attributes = {a: vehicle[a] for a in network.attributes("TRANSIT_VEHICLE")}
-            for line in lines[veh_id]:
-                line.vehicle = temp_veh
-            network.delete_transit_vehicle(vehicle)
-            new_veh = network.create_transit_vehicle(veh_id, meta_mode.id)
-            for a, v in attributes.items():
-                new_veh[a] = v
-            for line in lines[veh_id]:
-                line.vehicle = new_veh
-            network.delete_transit_vehicle(temp_veh)
+            if vehicle.mode.id != "p": #remove PNR dummy route from transit modes
+                temp_veh = network.create_transit_vehicle(get_vehicle_id(), vehicle.mode.id)
+                veh_id = vehicle.id
+                attributes = {a: vehicle[a] for a in network.attributes("TRANSIT_VEHICLE")}
+                for line in lines[veh_id]:
+                    line.vehicle = temp_veh
+                network.delete_transit_vehicle(vehicle)
+                new_veh = network.create_transit_vehicle(veh_id, meta_mode.id)
+                for a, v in attributes.items():
+                    new_veh[a] = v
+                for line in lines[veh_id]:
+                    line.vehicle = new_veh
+                network.delete_transit_vehicle(temp_veh)
         for link in network.links():
             link.modes -= transit_modes
         for mode in transit_modes:
